refactor: Move components to relevant directories
