+++ src/screens/TripDetails/components/CancelledDepartureMessage.tsx
