+++ src/screens/TripDetails/components/PaginatedDetailsHeader.tsx
